Split operations into tsp and websockets
Moving back closer to re-ducks:
https://medium.freecodecamp.org/scaling-your-redux-app-with-ducks-6115955638be
